+++ realm/realm-library/src/androidTest/java/io/realm/RealmModelTests.java
@@ -101,8 +101,7 @@ public class RealmModelTests {
         for (int i = 1; i < 43; i++) { // using i = 0 as PK will crash subsequent createObject
                                        // since createObject uses default values
             realm.beginTransaction();
-            AllTypesRealmModel allTypesRealmModel = realm.createObject(AllTypesRealmModel.class);
-            allTypesRealmModel.columnLong = i;
+            realm.createObject(AllTypesRealmModel.class, i);
             realm.commitTransaction();
         }